If a host goes down, do not start recovery operations
We have SLA's from h/w team, where if a entire host goes offline,
we do not trigger recovery/migration of data for pg's present in
all those osds of the host.

With the option "mon osd down out subtree limit = host", we are
enabling it, so that osds of the hosts are marked "down",
but not "out"

Note, if a individual osds go down, they get marked as down and out.

Signed-off-by: shishir gowda <shishir.gowda@ril.com>