Cinder volume active passive failover issue
when a node is in backup state, we dont want the c-vol service to be
running and we stop it. But this causes a side effect, wherein the node
switched to fault state as the keepalived health check script failed.
This side effect prevents the node from auto-transition to master state
during failover. so i've disabled the health check when the node is in
backup state.

Closes-Bug: #JBS-80